Transformation: Fix for internal routine extraction with literal kinds

+++ loki/transformations/extract/tests/test_extract_internal.py
@@ -336,7 +336,7 @@ def test_extract_internal_procedures_kind_resolution(frontend):
             contains
             subroutine inner()
                 integer(kind = jpim) :: y
-                integer :: z
+                integer(kind=8) :: z
                 z = y
             end subroutine inner
         end subroutine outer
